+++ NEWS.md
@@ -34,6 +34,21 @@ Outstanding ones only.
       You need to use a Hash literal to set a Hash to a first member.
       [[Feature #16806]]
 
+* Queue
+
+    * Queue#initialize now accepts an Enumerable of initial values.
+      [[Feature #17327]]
+
+* Thread::Backtrace
+
+    * Thread::Backtrace.limit, which returns the value to limit backtrace
+      length set by `--backtracse-limit` command line option, is added.
+      [[Feature #17479]]
+
+* $LOAD_PATH
+
+    * $LOAD_PATH.resolve_feature_path does not raise. [[Feature #16043]]
+
 ## Stdlib updates
 
 Outstanding ones only.
@@ -65,6 +80,9 @@ Excluding feature bug fixes.
 
 
 [Feature #14256]: https://bugs.ruby-lang.org/issues/14256
+[Feature #16043]: https://bugs.ruby-lang.org/issues/16043
 [Feature #16806]: https://bugs.ruby-lang.org/issues/16806
 [Feature #17312]: https://bugs.ruby-lang.org/issues/17312
+[Feature #17327]: https://bugs.ruby-lang.org/issues/17327
 [Bug #17423]: https://bugs.ruby-lang.org/issues/17423
+[Feature #17479]: https://bugs.ruby-lang.org/issues/17479